Make the "hts" parameter implicit
This makes HTTPSSVC more valuable as a security measure, for
websites that can't easily get themselves on the HSTS preload
list.

The value of the "_http" scheme prefix is unclear in this
formulation.  It could possibly be removed, depending on
the importance of supporting strange situations like
"https://example.com:80", or http and https on the same port.