Replace "owner name" with "query name"
There are two section headers where we imply that we are talking about
the "owner name" of a SVCB record, but this is not actually true due to
CNAMEs, aliases, etc.  This change uses "query name" and "owner name" in
a more consistent way, and clarifies the relationship with Attrleaf.

Fixes #250