Proposal: Simplify SvcParamKey registration policy
This change removes the First Come First Served registration policy
and the Expert Review policy, and combines them into a single
Specification Required range, with guidelines for what constitutes a
minimal specification.  The goal of this change is to avoid zone file
incompatibilities that could arise from the registration of keys whose
presentation format is poorly documented.