Relax the "no recourse" policy for HTTPS-RR-HSTS
This change adjusts the requirements to require termination without
recourse only on connections where the client can be reasonably sure
that it is talking to the actual server, not a middlebox.  This reflects
the view that the server can and should vouch for its own correct
configuration, but the client is better positioned to judge how to
handle middlebox-related failures.

Fixes #87